chore: transciber service (WIP)
